+++ main.go
@@ -2,11 +2,9 @@ package main
 
 import (
 	"os"
-	"runtime"
 
 	"github.com/Sirupsen/logrus"
 	"github.com/codegangsta/cli"
-	"github.com/opencontainers/runc/libcontainer"
 )
 
 const (
@@ -26,18 +24,6 @@ After creating a spec for your root filesystem with runc, you can execute a simp
     runc
 `
 )
-
-func init() {
-	if len(os.Args) > 1 && os.Args[1] == "init" {
-		runtime.GOMAXPROCS(1)
-		runtime.LockOSThread()
-		factory, _ := libcontainer.New("")
-		if err := factory.StartInitialization(); err != nil {
-			fatal(err)
-		}
-		panic("--this line should never been executed, congratulations--")
-	}
-}
 
 func main() {
 	app := cli.NewApp()
@@ -77,23 +63,8 @@ func main() {
 		}
 		return nil
 	}
-	// default action is to execute a container
-	app.Action = func(context *cli.Context) {
-		spec, err := loadSpec(context.Args().First())
-		if err != nil {
-			fatal(err)
-		}
-		if os.Geteuid() != 0 {
-			logrus.Fatal("runc should be run as root")
-		}
-		status, err := execContainer(context, spec)
-		if err != nil {
-			logrus.Fatalf("Container start failed: %v", err)
-		}
-		// exit with the container's exit status so any external supervisor is
-		// notified of the exit with the correct exit status.
-		os.Exit(status)
-	}
+
+	app.Action = runAction
 
 	if err := app.Run(os.Args); err != nil {
 		logrus.Fatal(err)
