+++ restore.go
@@ -1,3 +1,5 @@
+// +build linux
+
 package main
 
 import (
